Move purity checker to validate.rs
