+++ creusot/src/validate.rs
@@ -1,13 +1,19 @@
-use rustc_hir::{def::DefKind, def_id::DefId};
-use rustc_middle::ty::Visibility;
+use rustc_hir::{
+    def::DefKind,
+    def_id::{DefId, LocalDefId},
+};
+use rustc_middle::{
+    thir::{self, ClosureExpr, ExprKind, Thir},
+    ty::{FnDef, TyCtxt, Visibility},
+};
 use rustc_span::Span;
 
 use crate::{
     ctx::{parent_module, TranslationCtx},
-    translation::{
-        pearlite::{super_visit_term, TermKind, TermVisitor},
-        specification::is_overloaded_item,
-    },
+    error::{Error, InternalError},
+    pearlite::{pearlite_stub, Stub},
+    specification::contract_of,
+    translation::pearlite::{super_visit_term, TermKind, TermVisitor},
     util::{self, is_law, is_open_inv_result},
 };
 
@@ -110,6 +116,22 @@ pub(crate) fn validate_traits(ctx: &mut TranslationCtx) {
     }
 }
 
+// These methods are allowed to cheat the purity restrictions because they are lang items we cannot redefine
+fn is_overloaded_item(tcx: TyCtxt, def_id: DefId) -> bool {
+    let def_path = tcx.def_path_str(def_id);
+
+    def_path.ends_with("::ops::Mul::mul")
+        || def_path.ends_with("::ops::Add::add")
+        || def_path.ends_with("::ops::Sub::sub")
+        || def_path.ends_with("::ops::Div::div")
+        || def_path.ends_with("::ops::Rem::rem")
+        || def_path.ends_with("::ops::Neg::neg")
+        || def_path.ends_with("::boxed::Box::<T>::new")
+        || def_path.ends_with("::ops::Deref::deref")
+        || def_path.ends_with("::ops::DerefMut::deref_mut")
+        || def_path.ends_with("Snapshot::<T>::from_fn")
+}
+
 pub(crate) fn validate_impls(ctx: &mut TranslationCtx) {
     for impl_id in ctx.all_local_trait_impls(()).values().flat_map(|i| i.iter()) {
         if !matches!(ctx.def_kind(*impl_id), DefKind::Impl { .. }) {
@@ -206,5 +228,161 @@ pub(crate) fn validate_impls(ctx: &mut TranslationCtx) {
                 }
             }
         }
+    }
+}
+
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+pub(crate) enum Purity {
+    Program { terminates: bool, no_panic: bool },
+    Logic { prophetic: bool },
+}
+
+impl Purity {
+    pub(crate) fn of_def_id<'tcx>(ctx: &mut TranslationCtx<'tcx>, def_id: DefId) -> Self {
+        let is_snapshot = util::is_snapshot_closure(ctx.tcx, def_id);
+        if util::is_predicate(ctx.tcx, def_id) && util::is_prophetic(ctx.tcx, def_id)
+            || util::is_logic(ctx.tcx, def_id) && util::is_prophetic(ctx.tcx, def_id)
+            || util::is_spec(ctx.tcx, def_id) && !is_snapshot
+        {
+            Purity::Logic { prophetic: true }
+        } else if util::is_predicate(ctx.tcx, def_id)
+            || util::is_logic(ctx.tcx, def_id)
+            || is_snapshot
+        {
+            Purity::Logic { prophetic: false }
+        } else {
+            let contract = contract_of(ctx, def_id);
+            let terminates = contract.terminates;
+            let no_panic = contract.no_panic;
+            Purity::Program { terminates, no_panic }
+        }
+    }
+
+    fn can_call(self, other: Purity) -> bool {
+        match (self, other) {
+            (Purity::Logic { prophetic: true }, Purity::Logic { prophetic: false }) => true,
+            (
+                Purity::Program { no_panic, terminates },
+                Purity::Program { no_panic: no_panic2, terminates: terminates2 },
+            ) => no_panic <= no_panic2 && terminates <= terminates2,
+            (ctx, call) => ctx == call,
+        }
+    }
+
+    fn as_str(&self) -> &'static str {
+        match self {
+            Purity::Program { terminates, no_panic } => match (*terminates, *no_panic) {
+                (true, true) => "program (pure)",
+                (true, false) => "program (terminates)",
+                (false, true) => "program (no panic)",
+                (false, false) => "program",
+            },
+            Purity::Logic { prophetic: false } => "logic",
+            Purity::Logic { prophetic: true } => "prophetic logic",
+        }
+    }
+}
+
+pub(crate) fn validate_purity(ctx: &mut TranslationCtx, def_id: LocalDefId) {
+    let (thir, expr) = ctx
+        .thir_body(def_id)
+        .unwrap_or_else(|_| Error::from(InternalError("Cannot fetch THIR body")).emit(ctx.tcx));
+    let thir = thir.borrow();
+    if thir.exprs.is_empty() {
+        Error::new(ctx.def_span(def_id), "type checking failed").emit(ctx.tcx);
+    }
+
+    let def_id = def_id.to_def_id();
+    let purity = Purity::of_def_id(ctx, def_id);
+    if matches!(purity, Purity::Program { .. }) && crate::util::is_no_translate(ctx.tcx, def_id) {
+        return;
+    }
+
+    thir::visit::walk_expr(&mut PurityVisitor { ctx, thir: &thir, context: purity }, &thir[expr]);
+}
+
+pub(crate) struct PurityVisitor<'a, 'tcx> {
+    pub(crate) ctx: &'a mut TranslationCtx<'tcx>,
+    pub(crate) thir: &'a Thir<'tcx>,
+    pub(crate) context: Purity,
+}
+
+impl<'a, 'tcx> PurityVisitor<'a, 'tcx> {
+    fn purity(&mut self, fun: thir::ExprId, func_did: DefId) -> Purity {
+        let stub = pearlite_stub(self.ctx.tcx, self.thir[fun].ty);
+
+        if matches!(stub, Some(Stub::Fin))
+            || util::is_predicate(self.ctx.tcx, func_did)
+                && util::is_prophetic(self.ctx.tcx, func_did)
+            || util::is_logic(self.ctx.tcx, func_did) && util::is_prophetic(self.ctx.tcx, func_did)
+        {
+            Purity::Logic { prophetic: true }
+        } else if util::is_predicate(self.ctx.tcx, func_did)
+            || util::is_logic(self.ctx.tcx, func_did)
+            || util::get_builtin(self.ctx.tcx, func_did).is_some()
+            || stub.is_some()
+        {
+            Purity::Logic { prophetic: false }
+        } else {
+            let contract = contract_of(self.ctx, func_did);
+            let terminates = contract.terminates;
+            let no_panic = contract.no_panic;
+            Purity::Program { terminates, no_panic }
+        }
+    }
+}
+
+impl<'a, 'tcx> thir::visit::Visitor<'a, 'tcx> for PurityVisitor<'a, 'tcx> {
+    fn thir(&self) -> &'a thir::Thir<'tcx> {
+        self.thir
+    }
+
+    fn visit_expr(&mut self, expr: &'a thir::Expr<'tcx>) {
+        match expr.kind {
+            ExprKind::Call { fun, .. } => {
+                // FIXME: like in detect_recursion (MIR visitor), we would need to specialize the trait functions.
+                if let &FnDef(func_did, _) = self.thir[fun].ty.kind() {
+                    let fn_purity = self.purity(fun, func_did);
+                    if !self.context.can_call(fn_purity)
+                        && !is_overloaded_item(self.ctx.tcx, func_did)
+                    {
+                        let (caller, callee) = match (self.context, fn_purity) {
+                            (Purity::Program { .. }, Purity::Program { .. })
+                            | (Purity::Logic { .. }, Purity::Logic { .. }) => {
+                                (self.context.as_str(), fn_purity.as_str())
+                            }
+                            (Purity::Program { .. }, Purity::Logic { .. }) => ("program", "logic"),
+                            (Purity::Logic { .. }, Purity::Program { .. }) => ("logic", "program"),
+                        };
+                        let msg = format!(
+                            "called {callee} function `{}` in {caller} context",
+                            self.ctx.def_path_str(func_did),
+                        );
+
+                        self.ctx.dcx().span_err(self.thir[fun].span, msg);
+                    }
+                } else if !matches!(self.context, Purity::Program { .. }) {
+                    // TODO Add a "code" back in
+                    self.ctx.dcx().span_fatal(expr.span, "non function call in logical context")
+                }
+            }
+            ExprKind::Closure(box ClosureExpr { closure_id, .. }) => {
+                if util::is_spec(self.ctx.tcx, closure_id.into()) {
+                    return;
+                }
+
+                let (thir, expr) = self.ctx.thir_body(closure_id).unwrap_or_else(|_| {
+                    Error::from(InternalError("Cannot fetch THIR body")).emit(self.ctx.tcx)
+                });
+                let thir = thir.borrow();
+
+                thir::visit::walk_expr(
+                    &mut PurityVisitor { ctx: self.ctx, thir: &thir, context: self.context },
+                    &thir[expr],
+                );
+            }
+            _ => {}
+        }
+        thir::visit::walk_expr(self, expr)
     }
 }
